requirements.txt
numpy
matplotlib
Pillow
tkintertable
rasterio

// pip install numpy matplotlib Pillow tkintertable rasterio